Use ValueStringBuilder in preference to StringBuilderCache
