+++ src/System.Private.CoreLib/shared/System/Collections/Generic/KeyValuePair.cs
@@ -21,24 +21,26 @@ namespace System.Collections.Generic
         /// </summary>
         internal static string PairToString(object? key, object? value)
         {
-            StringBuilder s = StringBuilderCache.Acquire();
+            Span<char> initialBuffer = stackalloc char[64];
+            var s = new ValueStringBuilder(initialBuffer);
+
             s.Append('[');
 
             if (key != null)
             {
-                s.Append(key);
+                s.Append(key.ToString());
             }
 
             s.Append(", ");
 
             if (value != null)
             {
-                s.Append(value);
+                s.Append(value.ToString());
             }
 
             s.Append(']');
 
-            return StringBuilderCache.GetStringAndRelease(s);
+            return s.ToString();
         }
     }
