+++ src/System.Private.CoreLib/shared/System/Runtime/Intrinsics/Vector256_1.cs
@@ -174,20 +174,21 @@ namespace System.Runtime.Intrinsics
             ThrowHelper.ThrowForUnsupportedVectorBaseType<T>();
 
             int lastElement = Count - 1;
-            StringBuilder sb = StringBuilderCache.Acquire();
+            Span<char> initialBuffer = stackalloc char[64];
+            var sb = new ValueStringBuilder(initialBuffer);
             CultureInfo invariant = CultureInfo.InvariantCulture;
 
             sb.Append('<');
             for (int i = 0; i < lastElement; i++)
             {
-                sb.Append(((IFormattable)this.GetElement(i)).ToString("G", invariant))
-                 .Append(',')
-                 .Append(' ');
+                sb.Append(((IFormattable)this.GetElement(i)).ToString("G", invariant));
+                sb.Append(',');
+                sb.Append(' ');
             }
-            sb.Append(((IFormattable)this.GetElement(lastElement)).ToString("G", invariant))
-             .Append('>');
+            sb.Append(((IFormattable)this.GetElement(lastElement)).ToString("G", invariant));
+            sb.Append('>');
 
-            return StringBuilderCache.GetStringAndRelease(sb);
+            return sb.ToString();
         }
     }
 }
