+++ src/System.Private.CoreLib/shared/System/TimeZoneInfo.StringSerializer.cs
@@ -42,20 +42,21 @@ namespace System
             /// </summary>
             public static string GetSerializedString(TimeZoneInfo zone)
             {
-                StringBuilder serializedText = StringBuilderCache.Acquire();
+                Span<char> initialBuffer = stackalloc char[InitialCapacityForString];
+                var serializedText = new ValueStringBuilder(initialBuffer);
 
                 //
                 // <_id>;<_baseUtcOffset>;<_displayName>;<_standardDisplayName>;<_daylightDispayName>
                 //
-                SerializeSubstitute(zone.Id, serializedText);
+                SerializeSubstitute(zone.Id, ref serializedText);
                 serializedText.Append(Sep);
                 serializedText.AppendSpanFormattable(zone.BaseUtcOffset.TotalMinutes, format: default, CultureInfo.InvariantCulture);
                 serializedText.Append(Sep);
-                SerializeSubstitute(zone.DisplayName, serializedText);
+                SerializeSubstitute(zone.DisplayName, ref serializedText);
                 serializedText.Append(Sep);
-                SerializeSubstitute(zone.StandardName, serializedText);
+                SerializeSubstitute(zone.StandardName, ref serializedText);
                 serializedText.Append(Sep);
-                SerializeSubstitute(zone.DaylightName, serializedText);
+                SerializeSubstitute(zone.DaylightName, ref serializedText);
                 serializedText.Append(Sep);
 
                 AdjustmentRule[] rules = zone.GetAdjustmentRules();
@@ -69,9 +70,9 @@ namespace System
                     serializedText.AppendSpanFormattable(rule.DaylightDelta.TotalMinutes, format: default, CultureInfo.InvariantCulture);
                     serializedText.Append(Sep);
                     // serialize the TransitionTime's
-                    SerializeTransitionTime(rule.DaylightTransitionStart, serializedText);
+                    SerializeTransitionTime(rule.DaylightTransitionStart, ref serializedText);
                     serializedText.Append(Sep);
-                    SerializeTransitionTime(rule.DaylightTransitionEnd, serializedText);
+                    SerializeTransitionTime(rule.DaylightTransitionEnd, ref serializedText);
                     serializedText.Append(Sep);
                     if (rule.BaseUtcOffsetDelta != TimeSpan.Zero)
                     {
@@ -89,7 +90,7 @@ namespace System
                 }
                 serializedText.Append(Sep);
 
-                return StringBuilderCache.GetStringAndRelease(serializedText);
+                return serializedText.ToString();
             }
 
             /// <summary>
@@ -135,7 +136,7 @@ namespace System
             /// "]" -> "\]"
             /// "\" -> "\\"
             /// </summary>
-            private static void SerializeSubstitute(string text, StringBuilder serializedText)
+            private static void SerializeSubstitute(string text, ref ValueStringBuilder serializedText)
             {
                 foreach (char c in text)
                 {
@@ -150,7 +151,7 @@ namespace System
             /// <summary>
             /// Helper method to serialize a TimeZoneInfo.TransitionTime object.
             /// </summary>
-            private static void SerializeTransitionTime(TransitionTime time, StringBuilder serializedText)
+            private static void SerializeTransitionTime(TransitionTime time, ref ValueStringBuilder serializedText)
             {
                 serializedText.Append(Lhs);
                 serializedText.Append(time.IsFixedDateRule ? '1' : '0');
@@ -265,7 +266,8 @@ namespace System
                     throw new SerializationException(SR.Serialization_InvalidData);
                 }
                 State tokenState = State.NotEscaped;
-                StringBuilder token = StringBuilderCache.Acquire(InitialCapacityForString);
+                Span<char> initialBuffer = stackalloc char[InitialCapacityForString];
+                var token = new ValueStringBuilder(initialBuffer);
 
                 // walk the serialized text, building up the token as we go...
                 for (int i = _currentTokenStartIndex; i < _serializedText.Length; i++)
@@ -302,7 +304,7 @@ namespace System
                                 {
                                     _state = State.StartOfToken;
                                 }
-                                return StringBuilderCache.GetStringAndRelease(token);
+                                return token.ToString();
 
                             case '\0':
                                 // invalid character
